requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.0
redis==5.0.1
aioredis==2.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
pytest==7.4.4
pytest-asyncio==0.23.3
